Seventeenth commit. Restructured ID properties for Note and Category classes to differentiate between them. Crated GetNotesByCategory() method in NoteService class and corresponding Get() method in NoteController. ElevenNote 5.00. Will test GET/POST/PUT/DELETE methods in Postman before next commit. Now playing: Too Many Puppies by Primus

+++ ElevenNote.WebAPI/ElevenNote.Models/CategoryDetail.cs
@@ -9,7 +9,7 @@ namespace ElevenNote.Models
 {
     public class CategoryDetail
     {
-        public int CategoryID { get; set; }
+        public Int16 CategoryID { get; set; }
         public string Name { get; set; }
         public List<Note> Notes { get; set; }
         public DateTimeOffset CreatedUTC { get; set; }
